# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "description": "Stable structured error detail for API clients.",
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message_key": {
            "minLength": 1,
            "title": "Message Key",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message_key"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "ApiErrorResponse": {
        "additionalProperties": false,
        "description": "Stable error envelope returned by the API boundary.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "ProjectCompareRequest": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "minLength": 1,
            "title": "Analysis Id",
            "type": "string"
          },
          "left_scenario_id": {
            "minLength": 1,
            "title": "Left Scenario Id",
            "type": "string"
          },
          "project": {
            "additionalProperties": true,
            "title": "Project",
            "type": "object"
          },
          "right_scenario_id": {
            "minLength": 1,
            "title": "Right Scenario Id",
            "type": "string"
          }
        },
        "required": [
          "project",
          "analysis_id",
          "left_scenario_id",
          "right_scenario_id"
        ],
        "title": "ProjectCompareRequest",
        "type": "object"
      },
      "ProjectCompareResponse": {
        "additionalProperties": false,
        "properties": {
          "comparison": {
            "additionalProperties": true,
            "title": "Comparison",
            "type": "object"
          }
        },
        "required": [
          "comparison"
        ],
        "title": "ProjectCompareResponse",
        "type": "object"
      },
      "ProjectDocumentRequest": {
        "additionalProperties": false,
        "description": "Generic Project v2 document carried by a file-oriented client.",
        "properties": {
          "project": {
            "additionalProperties": true,
            "title": "Project",
            "type": "object"
          }
        },
        "required": [
          "project"
        ],
        "title": "ProjectDocumentRequest",
        "type": "object"
      },
      "ProjectDuplicateScenarioRequest": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "minLength": 1,
            "title": "Analysis Id",
            "type": "string"
          },
          "project": {
            "additionalProperties": true,
            "title": "Project",
            "type": "object"
          },
          "scenario_id": {
            "minLength": 1,
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_name": {
            "minLength": 1,
            "title": "Scenario Name",
            "type": "string"
          }
        },
        "required": [
          "project",
          "analysis_id",
          "scenario_id",
          "scenario_name"
        ],
        "title": "ProjectDuplicateScenarioRequest",
        "type": "object"
      },
      "ProjectFromAnalysisRequest": {
        "additionalProperties": false,
        "properties": {
          "analysis_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Name"
          },
          "analysis_snapshot": {
            "additionalProperties": true,
            "title": "Analysis Snapshot",
            "type": "object"
          },
          "project_name": {
            "default": "Untitled HCM study",
            "minLength": 1,
            "title": "Project Name",
            "type": "string"
          },
          "scenario_name": {
            "default": "Base",
            "minLength": 1,
            "title": "Scenario Name",
            "type": "string"
          }
        },
        "required": [
          "analysis_snapshot"
        ],
        "title": "ProjectFromAnalysisRequest",
        "type": "object"
      },
      "ProjectRecordResultRequest": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "minLength": 1,
            "title": "Analysis Id",
            "type": "string"
          },
          "analysis_snapshot": {
            "additionalProperties": true,
            "title": "Analysis Snapshot",
            "type": "object"
          },
          "project": {
            "additionalProperties": true,
            "title": "Project",
            "type": "object"
          },
          "scenario_id": {
            "minLength": 1,
            "title": "Scenario Id",
            "type": "string"
          }
        },
        "required": [
          "project",
          "analysis_id",
          "scenario_id",
          "analysis_snapshot"
        ],
        "title": "ProjectRecordResultRequest",
        "type": "object"
      },
      "ProjectRenameScenarioRequest": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "minLength": 1,
            "title": "Analysis Id",
            "type": "string"
          },
          "project": {
            "additionalProperties": true,
            "title": "Project",
            "type": "object"
          },
          "scenario_id": {
            "minLength": 1,
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_name": {
            "minLength": 1,
            "title": "Scenario Name",
            "type": "string"
          }
        },
        "required": [
          "project",
          "analysis_id",
          "scenario_id",
          "scenario_name"
        ],
        "title": "ProjectRenameScenarioRequest",
        "type": "object"
      },
      "ProjectResponse": {
        "additionalProperties": false,
        "properties": {
          "migrated": {
            "default": false,
            "title": "Migrated",
            "type": "boolean"
          },
          "project": {
            "additionalProperties": true,
            "title": "Project",
            "type": "object"
          }
        },
        "required": [
          "project"
        ],
        "title": "ProjectResponse",
        "type": "object"
      },
      "ProjectUpdateScenarioRequest": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "minLength": 1,
            "title": "Analysis Id",
            "type": "string"
          },
          "analysis_snapshot": {
            "additionalProperties": true,
            "title": "Analysis Snapshot",
            "type": "object"
          },
          "project": {
            "additionalProperties": true,
            "title": "Project",
            "type": "object"
          },
          "scenario_id": {
            "minLength": 1,
            "title": "Scenario Id",
            "type": "string"
          }
        },
        "required": [
          "project",
          "analysis_id",
          "scenario_id",
          "analysis_snapshot"
        ],
        "title": "ProjectUpdateScenarioRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local application boundary for the rebuilt HCM workspace.",
    "title": "HCM Calculator API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/compare": {
      "post": {
        "operationId": "project_compare_api_v1_projects_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCompareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Project Compare",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/duplicate-scenario": {
      "post": {
        "operationId": "project_duplicate_scenario_api_v1_projects_duplicate_scenario_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDuplicateScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Project Duplicate Scenario",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/from-analysis": {
      "post": {
        "operationId": "project_from_analysis_api_v1_projects_from_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectFromAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Project From Analysis",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/record-result": {
      "post": {
        "operationId": "project_record_result_api_v1_projects_record_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRecordResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Project Record Result",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/rename-scenario": {
      "post": {
        "operationId": "project_rename_scenario_api_v1_projects_rename_scenario_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRenameScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Project Rename Scenario",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/update-scenario": {
      "post": {
        "operationId": "project_update_scenario_api_v1_projects_update_scenario_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdateScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Project Update Scenario",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/validate": {
      "post": {
        "operationId": "validate_project_api_v1_projects_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Validate Project",
        "tags": [
          "projects"
        ]
      }
    }
  }
}